+++ ESP32_PianoLux/FadingRunEffect.ino
@@ -1,6 +1,6 @@
 //FadingRunEffect.cpp
 #include "FadingRunEffect.h"
-FadingRunEffect ::FadingRunEffect(int effectLen, int startPosition, CHSV splashColor, int headFadeRate, int velocity) {
+FadingRunEffect ::FadingRunEffect(uint8_t effectLen, uint8_t startPosition, CHSV splashColor, uint8_t headFadeRate, uint8_t velocity) {
   // Initialize the effect parameters
   this->effectLen = effectLen;
   this->startPosition = startPosition;
@@ -11,24 +11,24 @@ FadingRunEffect ::FadingRunEffect(int effectLen, int startPosition, CHSV splashC
   this->velocity = velocity;
 }
 
-int MAX_VALUE = 255;
-int LOWEST_BRIGHTNESS = 50;
-int LOWEST_SATURATION = 200;
+uint8_t MAX_VALUE = 255;
+uint8_t LOWEST_BRIGHTNESS = 50;
+uint8_t LOWEST_SATURATION = 200;
 
-int HEAD_FADE_LOW_THRESHOLD = 50;
-int HEAD_FADE_HI_THRESHOLD = 255;
+uint8_t HEAD_FADE_LOW_THRESHOLD = 50;
+uint8_t HEAD_FADE_HI_THRESHOLD = 255;
 
-int FadingRunEffect ::getSaturation(int velocity) {
+uint8_t FadingRunEffect ::getSaturation(uint8_t velocity) {
   return adjustValue(velocity, LOWEST_SATURATION, MAX_VALUE);
 }
 
-int FadingRunEffect ::getBrightness(int velocity) {
+uint8_t FadingRunEffect ::getBrightness(uint8_t velocity) {
   return adjustValue(velocity, LOWEST_BRIGHTNESS, MAX_VALUE);
 }
 
-int FadingRunEffect ::calcOffset(int step, int velocity) {
-  int adjustedVelocity = adjustValue(velocity, HEAD_FADE_LOW_THRESHOLD, HEAD_FADE_HI_THRESHOLD);
-  int offset = step * adjustedVelocity / MAX_VELOCITY;
+uint8_t FadingRunEffect ::calcOffset(uint8_t step, uint8_t velocity) {
+  uint8_t adjustedVelocity = adjustValue(velocity, HEAD_FADE_LOW_THRESHOLD, HEAD_FADE_HI_THRESHOLD);
+  uint8_t offset = step * adjustedVelocity / MAX_VELOCITY;
   if (offset > step) {
     return step;
   } else {
@@ -37,16 +37,16 @@ int FadingRunEffect ::calcOffset(int step, int velocity) {
 }
 
 
-int FadingRunEffect ::adjustValue(int value, int lowerThreshold, int maxValue) {
+uint8_t FadingRunEffect ::adjustValue(uint8_t value, uint8_t lowerThreshold, uint8_t maxValue) {
   return lowerThreshold + (value * (maxValue - lowerThreshold) / maxValue);
 }
 
-void FadingRunEffect ::setHeadLED(int step) {
+void FadingRunEffect ::setHeadLED(uint8_t step) {
   if (step > effectLen) {
     return;
   }
-  int pos1 = ledNum(this->startPosition + calcOffset(step, velocity));
-  int pos2 = ledNum(this->startPosition - calcOffset(step, velocity));
+  uint8_t pos1 = ledNum(this->startPosition + calcOffset(step, velocity));
+  uint8_t pos2 = ledNum(this->startPosition - calcOffset(step, velocity));
   if (isOnStrip(pos1)) {
     if (splashColor != CHSV(0, 0, 0)) {
       leds[pos1] = CHSV(splashColor.hue, splashColor.saturation, getBrightness(velocity));
@@ -71,7 +71,7 @@ void FadingRunEffect ::nextStep() {
   setHeadLED(this->step);
   this->step++;
 }
-int FadingRunEffect ::getSteps() {
+uint8_t FadingRunEffect ::getSteps() {
   return effectLen;
 }
